Pine Needles - Gadgil & White Rot & DIC PPL update
The one rjags file is adjusted so all data is present for the 5 models. Now we have DIC and PPL criteria. The rjags needs to be edited depending on the incubation set (A 2017-2018 or B 2018-2019) and model. Next steps will be to try the WAIC instead of the DIC, since it is more commonly used.